Add API to change editor selection to given TextRange
This API should be used sparingly, setSelection() is still preferred
because it remembers the cursor locations for e.g. linewise visual mode.

The getNativeSelection() API is similarly only meant to be used to check
if a non-zero-length selection is present. If a Vrapper selection is
active, getNativeSelection() will return a constant which cannot be
queried without triggering an exception. If you want to know more, check
for the constant and call getSelection().