Rework sort to extend AbstractLineWiseOperation
This makes sure that all line-range calculations are centralized in
SimpleLineRange or other implementations.

In this case, the SortOperation had a somewhat-off calculation of the
'length' parameter, meaning that the original length of plain visual
selections was being used rather than the length of text between the
beginning of the first line and the end of the last line.

Fixes #644.