Reset commandline colors when it is opened
Changing Eclipse's style can sometimes reset widgets automatically and
sometimes can't. On Ubuntu, somehow the background color would get reset
but the foreground color wouldn't, ending up with black text on a black
background.

It's also not correct to reuse a Color instance like this, as this Color
may get disposed when colors are changed.

Fixes #680.