Pull LocalConfiguration creation into platform.
Moving it should allow us to hook in different mechanisms for grabbing
the default value of an Option, e.g. by looking at the Eclipse settings.

This move also has the advantage that platform services for the current
editor (e.g. search) can use the local configuration rather than the
global one so that they finally react to setlocal.