Introduce DefaultConfigProvider, integrate
SimpleConfiguration will now ask all DefaultConfigProviders to supply a
proper default for an Option. This should allow the platform to hook in
the appropriate dynamic values.

Also refactor SimpleLocalConfiguration to reuse as much as possible from
SimpleConfiguration. The whole 'get' logic can now also be handled by a
DefaultConfigProvider delegating to the parent configuration.