Implement vrapper-sneak Mk. II
Add a motion which will retrieve the currently cached sneak state and
jumps to the next location.

State is kept inside the SneakInputMode which should be created for each
editor. This mode will also accept input for interactive sneak usage.

It is necessary to bind `<Plug>(sneak-next)` to `;` before sneak can be
repeated quickly. This will still allow to use the last fX motion but
will not pick up on dfX like Vim does. This can be handy to sneak around
and use dfx using the repeat-command (the dot key).

The SneakStateManager is a helper interface which
EditorAdaptorStateManagder implements, this way tests don't need to know
where Sneak's cache / state is kept; they can just provide their own
stub which returns a correctly intialized state.

SneakMotion is not a jumping motion, especially not when repeating it
using `<Plug>(sneak-next)`. A decorator is used instead to pass off the
first invocation of SneakMotion as a jump so that MotionCommand stores
our previous location. After that MotionCommand will request an adaptor
to NavigatingMotion which will unwrap the decorator, this way repeating
won't add to the jump list again.

Note that SneakInputMode will still keep a global copy of the
SneakMotion even though we have a reserved spot for NavigatingMotions in
RegisterManager. This copy is meant for when the user does `s<CR>` to
reuse the last sneak search keyword without the risk of some other
plugin clobbering the last instance in RegisterManager.