Reset caret size each time setCaret is called
Previously the code was tweaked so as not to create and dispose a Caret
widget each time setCaret is called (usually once per command).

Eclipse does occasionally change the caret as well without telling
Vrapper. This means that the caret would become stuck as a small line
even though the block caret from normal mode should be shown.

Now the size of the caret is cached and its size (and hence width) is
reset each time setCaret is called. Even if the caret is changed by
Eclipse, it will be reset by just a single movement.

Note that the cached size can be off if the user changes the font size.
This can be easily worked around by switching to insert mode or
temporarily disabling Vrapper.