Introduce VisualCaretPainter, remove caret hacks
Vim puts the caret at the end of the line when doing a whole-line
inclusive selection. Eclipse on the other hand uses exclusive selection
so that the caret would be on the next line in column 0.

A previous workaround was to visually shrink the selection while Vrapper
kept the actual selected text internally. This new fix introduces a
PaintListener which will move the caret to its supposed position without
touching the selection.

The advantage is that the selection background of the last line now
stretches to the far right instead of being cut short at the last
character of the line. The caret can also be displayed within the
selection when using visual line mode, this makes it more clear why
pressing <Right> won't immediately select the lext line. Finally we can
also for a caret to be visible in blockwise mode as Eclipse hides it by
default.