Blockwise selection is inclusive in Vrapper
Horizontal caret swapping showed an off-by one behaviour: making a block
selection of two characters wide and pressing `O` would never swap the
caret. This was because the selection range was calculated as exclusive,
so the range's length would be 1 when there were actually 2 chars
selected.

Changing the range to inclusive means the selection's `from` and `to`
positions need to be used to do a proper swap. There was also some off-
by-one problem when running a change operation.