Fetch ITextViewer's IDocument for every search.
Caching the IDocument instance inside of a FindReplaceDocumentAdapter
means that this copy can get out of date.

One such case was in an editor which initialized to an IDocument
containing the string "Loading". When the editor would replace the
IDocument with actual contents, Vrapper's search would still use the old
cached value.

Work for #650.