Fix the dreaded "smart insert moves caret back" bug
It turns out that doing textWidget.replaceTextRange has subtle different
behaviour when Eclipse's "linked mode" is active:
- when not active, doing a zero-length replace (i.e. an insert) on the
text before the caret will shift the caret with it. This is handled by
the code in StyledText and is what the old code counted on:
it would place the caret on the next character or line, do the replace
and then move the caret a few characters back to be in the right spot.
- when linked mode is active, JFace will install some listeners on the
StyledText which will do the replace instead of StyledText. This
code will move the caret to *just after the insert*, fooling the code
which then steps back a few characters and ends up just *before* the
newly inserted character.

Inserting at the end of the file can still do a simple replace as we
know we should end up at the *new* end of the file. The fixed version
calls smartInsert (int, String) to do the right thing in all other
cases.

More explanation on the original weirdness is given by @keforbes in the
comments of issue #77.

Fixes #422.