Implement backtracking in remaps
Insert-mode allows us to type a continuous stream of characters, only
some of those could be meant for remapping. If we start looking at a
promissing match, we should also consider that longer spans might start
at the wrong point and must be scanned for other matches.

To give an example: let's say we have a remaping 'kj' instead of Escape.
Typing 'luckkj' should result in 'luck' and a return to Normal mode.
Before this fix we whould end up with 'luckkj' in the editor.

To fix this we ask the current mode if it allows for backtracking, if so
we deal with the initial matching key and recursively try to remap the
remaining characters. This recursion will at one point stop as we run
out of characters.

Fixes #669 and #677.