Fix visual block selection caret for first character
There was a check in place to make sure an edge-case visual offset can't
crash Vrapper, instead the caret painter would quit.

Sadly the edge condition was wrong, the character at offset zero *can*
be selected. Make the check for EOF a bit stricter so that the remaining
code works fine.