EXPERIMENTAL - Disable 'visualother' by default
Even thoulgh 'visualother' has been working quite stable for a while it
might not be what we want in the long run: it's all or nothing.

For example, opening Eclipse's Quick Outline and then pressing <Return>
on one of the selected members will create a selection for the name of
that chosen member. The 'visualother' feature will immediately move us
to visual mode and hence moving around will change the selection rather
than moving around freely.

On the other hand all Eclipse selections will automatically stick to
visual mode, e.g. Extend Selection, Shift Line(s) Up, ...  without extra
configuration or effort, constantly keeping the two in sync.

In the end the choice of setting 'visualother' is best left to the user.