Initialize nested editors which didn't exist at "Editor Opened" event
The Jasper Report editor appears to add a third "Preview" tab shortly
after the editor has been opened and activated. This triggered a
safeguard condition in Vrapper as it would think that the internal child
editorinfo got corrupted.

Simply detect the case and intercept nested editors during the
activation phase later on. In this case it wasn't even a useful editor
for Vrapper to operate on.

Fixes #727.