EXPERIMENTAL: Suppress platform event handlers during eclipse command
EclipseCommandMotion resets the selection as part of the actual Eclipse
movement commands. This in turn would trigger the SelectionVisualHandler
to switch to Normal mode which resulted in a "conflicted selection" as
Vrapper now displays a selection in Normal mode.

EclipseCommandHandler also sees these commands and might be duplicating
effort which is already happening in the motion code.

Regular motions work around this problem by setting the "selection in
progress" flag which SelectionVisualHandler will recognize. Set up a
similar flag for Eclipse commands and disable SelectionVisualHandler and
EclipseCommandHandler when this flag is active.

This fix was sorely needed to get eclipse-motion plugs working in visual
mode.