EXPERIMENTAL: Make sure visual Eclipse motions use the 'To' position
Eclipse motions are unaware of Vrapper, hence they will treat the end of
the editor's selection as the current position of the caret. This
behavior will act weird e.g. when pressing `End` followed by `Home` as
this would select characters on the *next* line. This is because the
editor's selection would actually extend until the next line.

Resetting the selection right before executing the Eclipse command and
then repairing it is the way to go.