EXPERIMENTAL: Allow Eclipse motions to move past end-of-line, fix caret
Normal mode would always move the caret back to be "on" the last
character of the line when an Eclipse command would move it to the end-
of-line (take "go to next word" for instance). Because the next
invocation of that Eclipse command would jump once more to the end-of-
line, you'd get an endless cycle of jumping back and forth.

Note that not stuff like Home or End should never jump across lines,
these are marked as special.