EXPERIMENTAL: Make other commands immediately leave visual mode
Executing commands through <Plug>(eclipse: keys would already make
Vrapper leave visual mode so it makes sense to also do this while the
user is working directly in Eclipse instead of executing them through a
remapping or macro.

This might surprise users of Cut / Copy / Paste who are used to the
current behavior but it *is* more consistent with Vim.