Catch and wrap exception when asking for line info of an invalid offset
Careless conversion of a Position instance to a view offset may yield a
negative offset because the Position happened to be on some folded text.

Throw a clearer exception indicating that it was a view offset, though
it could have been better if we had the model offset as well.