:let @x="" command will no longer clobber default register
Having a let @x="" command in your .vrapperrc would fill the x register
each time a new editor is created and clobber the default register at
the same time. This is annoying when opening a fresh editor after you
yanked something in the previous editor with the intention of pasting it
in the fresh one.

Call the freshly overloaded setContent(..) method to keep things
separate.