Allow macro to be appendeded to without triggering recursion detection
The recursion detection code would react immediately when somebody
requested a MacroPlayer for the same macro we've seen moments ago - in
this case when the user wants to run a macro repeatedly.

The fix is simply to record the current macro only when the MacroPlayer
starts executing its buffer.

Fixes #737.