Add :subwordmappings command to set up mappings
Users could bind everything themselves but it's just boilerplate. This
command will automate the bindings for them.

It works in 4 or 5 ways:

- running `:subwordmappings` with no parameters will require you to
press `<Leader>b` for the back motion, `i<Leader>` for the textobj.
- running `:subwordmappings ,` will assign the back motion to `,b`. The
textobject is triggered by `i,`.
- running `:subwordmappings , _` will assign the back motion to `,b` but
the textobject to `i_`.
- finally, the first parameter also accepts `''` to indicate that no
mapping prefix should be used, hence binding subword-back to `b`.
The (optional) second parameter works like before, if not present it
is set to `<Leader>`.

Related to #460.