NormalMode accepts multiple hints, find command with new utils method
A new plugin mode would pass in a hint for visual mode as the first item
in the hints list but Normal mode only checks for commands in the very
first position. Simply introduce an utility method which will scan all
hints to find the the first of a specific type.