Move extension mode initialization to getMode(x) call
Sneak mode will need a custom mode instance to store its state in the
current editor. However, the current editor might not have an instance
of that mode yet when it is freshly initialized. However, we might want
to try to invoke a SneakMotion stored in the lastNavigatingMotion
register yet that would raise a NullPointerException because the custom
mode is null.

Moving the extension loading code to getMode(x) means that retrieving
the mode will actually request the custom mode to be created,
immediately initializing its state and preventing null pointers in the
case of a vim-sneak port.