Paint "conflicted selection" caret inside of selection in normal mode
All platform methods returning the current Position will actually return
the offset of the last character in the selection so it makes sense to
paint the caret there.

Also make sure to keep painting it inside the selection when the caret
type changes. A freshly constructed Caret widget does not trigger a
repaint automatically and thus its initial location is not fixed by the
VisualCaretPainter.