Implement sneak text object
This text object is commonly bound to z / Z so as not to conflict with
surround.vim. It doesn't prompt for input like interactive sneak,
instead there is a state which will capture keys until a given threshold
is reached (default 2) or RETURN is pressed.

Note that this doesn't bind an alternative to d; or d, - those commands
will still use the Vrapper default ; motion, which is to use whatever fX
or dfX motion was executed last.