Keep unconsumed keys and original keys in separate collections
We want to keep the original input when a mapping succeeds to be able to
record a unique id of each mapping which is currently active. When a
recursive mapping got constructed, this id will be present on the stack
and hence infinite recursion can be avoided.

Using the unconsumed keys list for this purpose was not correct though
as mappings might exist which are prefixes of others (e.g. `\s` and
`\sss`). When the user then enters `\ssa` the result should be to run
mapping `\s` and type `sa`. However, conflating the original inputs and
unconsumed keys list would thus result in running `\s` and typing `\ssa`
as keys (potentially trigger endless recursion when we have backtracking
remap functionality like in Insert mode).

Keeping the two separate solves these problems.

Fixes #758.