Allow default vrapperrc location to be overriden
It might be interesting to use a different vrapperrc location when
testing a different version of vrapper or reproducing a problem.

This change will look for a system property called "vrapper.vrapperrc"
which will be used as location for the vrapperrc. This can be a
filename, relative to the user's homedir, or it can be an absolute
filepath.