Introduce NavigatingMotion, generalize ContinueFindingMotion
This interface can later be implemented by sneak or easymotion so that
the ContinueFindingMotion can launch the previous invocation stored in
the RegisterManager.

MotionCommand will check for an adapter to NavigationMotion each time a
motion is run through the GoThereState and, if available, store that
motion in the lastNavigatingMotion register. As a result the
lastNavigatingMotion register will contain the last FindCharMotion which
was invoked as `fX`, but it won't be clobbered when using dfX unlike the
default lastFindCharMotion register.

This way the enhanced ContinueFindingMotion could be used to mix fX and
dfX without continuously clobbering the last character.