Fix shifted caret by overloading all methods which Eclipse Neon might call
The shifted / tweaked caret we use for "input pending", "conflicted
selection" or "replace mode" would no longer be shown at the bottom of
the current character starting from Eclipse Neon.

As it turns out this Eclipse version started calling different methods
than just setLocation(int,int) to move the caret. This made it so that
the shifting code would no longer run and the caret would thus appear at
the top of the current character.

Simply overload all methods which can pass a position and point them to
our overload which shifts the offsets.