Fix NoSuchElementException in recursive mapping loop detection
Backtracking after a failed mapping match must handle every key
recursively to do a new mapping scan. Such a recursive invoke might in
turn trigger a mapping which then (correctly) gets stored in the mapping
stack to do loop detection.

The stack is popped, but when the recursive key handling returns from
all the backtracking then the `didMappingSucceed` function of
KeyStrokeTranslator might have a different state. This caused the
mapping stack to be "popped" when it was already empty.

Caching the value of `didMappingSucceed` before any recursion happens
and checking the contents of the stack should avoid this.

Fixes #793.