List all available global marks whether or not target file is opened
The getAllMarks implementation would only find global marks in currently
opened editors, but global marks are supposed to be persisted between
sessions and thus be listed irregardless if they are currently opened!

Simply copy the code of getGlobalMarker to run on the entire workspace,
this way we can find all the markers but without any need for the target
files to be open.

This problem got reported as part of #766.