Rework GoToMarkCommand to use proper Eclipse API, less looping
The former getGlobalMarkEditor code in EclipseCursorAndSelection code
would use the concrete WorkbenchPage class (generally better to use the
interface only) and looped over all open editors trying to find a mark.
This bit got reworked and was thus removed from this class.

GoToMarkCommand used to depend on Vrapper's FileService to open the file
associated with the mark, but it would do so based on *project-relative*
path. This would break when the current file and target file are in
different projects.

It was thus rewritten to start by looking up the global mark, find its
associated file on disk and then use plain Eclipse API to find an
already opened editor or make a fresh one for that file. Since we then
immediately get the Eclipse IEditorPart, looking up the target Vrapper
instance's EditorAdaptor becomes easier as well.

(Last but not least: normalized the whitespace in touched classes, fixed
warnings and removed suppress-warnings-annotation)

Fixes #766.