Vrapper-sneak: fix infinite looping when in column mode
Vrapper would be stuck in an infinite loop when the start or end of the
file contains a word matching the Sneak search string and it would not
be considered due to the column filter.

Regular sneak mode did not fail on it but it would still do too much
work.

Fixes #819.