Future-proofing modifier key detection, fix unknown 'special' key handling
Vrapper has always had a list of hardcoded key events whose keycodes it
should block: the modifier key presses which are received before those
modifiers are used in a shortcut (e.g. a Control key event is received
before a Control+R event). These key events have a 'character' field
value of 0x00, which would be treated as if it were an '@' keypress and
hence cause trouble down the line by triggering macros.

Since SWT now added the ALT_GR modifier in its newest version, this
scenario is occuring again. We cannot use the ALT_GR constant without
depending on the newest SWT version. This fix instead uses the
SWT.MODIFIER_MASK constant (which should actually have been used from
the start because the API guarantees it matches ALT, CONTROL, COMMAND or
whatever SWT adds in the future).

On top of that the logic for translating Control+Alpha keys was too lax.
While the key event's character value 0x00 can be used to indicate a
Ctrl+@ keypress, it is obviously also the default value for that field.
When Vrapper failed to detect a certain 'special key' it would in the
past have hit the Control+Alpha translation and hence interpret any key
with the character field value 0x00 as the '@' key. Any keypress
following it would then likely execute the associated macro or whatever
junk was in that register, hence corrupting the current document.

The new code will check that the Control key is actually held down and
that SWT's 'special key' bit is not set. As a bonus it will also log the
unknown key's event values if Vrapper's debug loggin is activated.

Work for #817