Do not fetch ITextEditor in mode constructor
All extension modes get constructed as soon as any extension mode is
requested.

Due to the way Vrapper-Sneak works Vrapper would eagerly construct the
Sneak mode and by extension the InsertExpandMode. This would sometimes
happen at a time when there's currently no Vrapper instance active or
when InsertExpandMode would get constructed for a Vrapper instance which
does not have focus (e.g. when toggling Vrapper). The result would be
that the initialization either fails or connects to the wrong editor.

Instead fetch the editor only when the mode gets used the first time,
from that point on the editor instance can be cached.