Add new Eclipse command/handler to grab key input and send it to Vrapper
In the past we would tell people to unbind common shortcuts in Eclipse
to make sure they're passed to Vrapper.

This new command will grab the key which was used to trigger it and will
send it to the currently focussed editor. If there's no editor focused
it will be ignored. Either way, bind it to one of the "Vrapper (Mode)
Active" contexts for maximum effect.

Fixes #108.