Add command/handler which sends out arrow key presses when triggered
We have had requests from people for easier ways to navigate the package
explorer, preferably using the HJKL keys like Vim does.

This commit adds a "Vrapper Up Arrow Map" command (and so on for the
other 3 arrows) to the General > Keys preference page which can be bound
to HJKL.

Do note that it should be bound to the "Vrapper Enabled on a View"
context to be safe; otherwise you would be unable to type HJKL in any
dialog or entry box.

As a safety measure the code will also fall back to just inserting the
bound keys if Vrapper is disabled or when the focused control is a
textbox. Other controls have not been tested because they're not
commonly available in Eclipse Views.

Fixes #173.