Rework "Vim's key bindings" scheme to map frequent shortcuts to Vrapper
The recent commits have given Vrapper a way to intercept any keyboard
shortcut it wants to, but people might have been accustomed to Vrapper
taking a backseat.

This commit thus leaves the default keyboard scheme mostly alone; it
only binds a sequence like "F15, Ctrl+A" as a sample binding so that
anyone using the default scheme can simply copy that command as a
starter and then change the sequence.

If people really want everything Vim-like they can now opt to select the
"Vim's key bindings" scheme and thus use all the Vrapper features
without any further setup. The downside is now that all useful Eclipse
shortcuts are suppresed, but then people can still copy that particular
shortcut and assign the copy to e.g. the "Vrapper Insert Mode Active"
context.

Final note: the sample "F15, Ctrl+A" shortcuts will remain visible even
when "Vim's key bindings" is selected. They are unlikely to be ever used
so it's not like they're in the way. It's only a bit messy.