Fix out-of-range error in End-Of-File motion
The motion was trying to jump to "before the last character in the file"
by subtracting 1 character, but on Windows this might put us in the
middle of a carriage return / line feed pair.

This commit changes it not to do any calculation, Vrapper is smart
enough to display the carret as usual on an empty last line. In fact,
the behaviour of 'G' (go to last line) also puts the caret on the last
empty line instead of on the last character of the last non-empty line.