Reset position changes in MoveUpDownNonWhitespace motion
First of all the code no longer uses MotionCommand.doIt(...).

MotionCommand does all kinds of thing like checking if the jump mark
needs to be updated, or if the caret is outside the window and hence
needs a scroll operation to make it visible again.

Since a composite motion should only have to do this *after* all
intermediate offsets have been traveled through it is far better to just
use the nested motion's "destination" method and set the intermediate
caret offset ourselves.

Then when we're finally done we should make sure to reset the caret to
the original offset so that calling destination(...) twice doesn't put
us at a different location each time.

Fixes #822.