Fix insertion code to move caret and do nothing for readonly text field
SWT StyledText doesn't seem to have a method of inserting text right
before the caret so that the caret is at the right place for more input.

Simply track our position based on the end of the text in the text field
and then move the caret to were it now should be.

Also make sure to keep read-only text fields untouched because the user
would then be unable to clear them.