Refactor motions to accept initial 'from' Position
Some Motion implementations call other motions. Some changed the caret
position to make those other motions return different destinations, but
forgetting to jump back to the initial position has side effects when
the motion is repeated in the same command.

Now a Motion accepts a Position to start from so that combined motions
never need to call setPosition(...) anymore. This refactoring does mean
that everything invoking motions needs to pass the current position, but
this might come in handy for our port of the Easymotion plugin.

Related to #822.