Fix IllegalArgumentException caused by returning -1 offset, normalize whitespace
The stickyColumnAtViewLine() method was using exceptions as flow control.
In there was a method throwing an exception to signal an error condition,
and that was replaced with a method returning -1 because the former was
deprecated. Suddenly the flow changed though because the -1 was not handled,
which goes to show why "exceptions as flow control" violates the "principle
of least astonishment" and hence is an anti-pattern.

Also replace tabs with spaces (it's the most common indenting style in this file)
and replace `if(...)` with `if (...)`.

Closes #875.